+++ EssentialFeed/Feed Cache/Infrastructure/CoreData/CoreDataFeedStoreModels/ManagedCache.swift
