+++ EssentialFeed/Feed Cache/Infrastructure/CoreData/CoreDataFeedStoreModels/ManagedFeedImage.swift
